Improve list sorting functionality
It’s possible to specify case-insensitive searches on a table in
list.js by adding a `data-insensitive=“true”` attribute to the buttons
in HTML. This feature is not documented and only became apparent in
reading the list.js source code.

Another improvement has been applied to the list.js code itself: the
addition of a regular expression that strips out the HTML tags from the
search query so that only content is sorted. See here:
https://github.com/javve/list.js/issues/303.

For the time being I’m going to rely on a custom “fork” of the list.js
code to be able to use this feature. Hopefully it will get merged into
the main codebase at some point.